feat(M6.10): добивка M6.5 квестов — персист, ?-иконки, item-collect
Закрывает хвост M6.5 (kill-флоу разблокирован фиксом M7 #11). Проверено клиентом.

Персист квестов (инкр.1):
- Таблица character_queststatus (alexwow_auth): owner+quest (PK), slot, status
  (0=активен/1=сдан), counter0..3. GetQuestStatusesAsync/UpsertQuestStatusAsync.
- Персист при accept/прогрессе/сдаче; загрузка LoadQuestStateAsync ДО спавна.
- Поля PLAYER_QUEST_LOG кладутся в НАЧАЛЬНЫЙ спавн игрока (PlayerSpawn) — иначе
  досылка апдейтом = «новое взятие» (звук+надпись) при релоге.

?-иконки приёмщиков (инкр.3):
- QuestStore хранит карты entry→questIds (giver/ender). StatusForAsync: приёмщик
  с выполненным квестом → REWARD(10, жёлтый ?); дающий с берущимся → AVAILABLE(8, !);
  в процессе → INCOMPLETE(5, серый ?). Значения сверены с TrinityCore QuestDef.h.
  ! пропадает после взятия/сдачи.

item-collect цели (инкр.4):
- QuestProgress.ReqItem[]/ReqItemCount[] + ObjectivesMet (существа+предметы).
- Квест-предметы теперь падают с мобов: GetCreatureLootAsync берёт и
  ChanceOrQuestChance<0; LootHandlers роняет их только держателю квеста
  (NeedsQuestItem) по |chance|%. Хук OnItemGainedAsync (InventoryGrant) → зачёт.
- Списание квест-предметов при сдаче (ConsumeItemsAsync) — освобождает слоты под
  награду.

Фикс окна сдачи: SMSG_QUESTGIVER_OFFER_REWARD поле auto_finish — u8, не u32
(сверено с TC QuestGiverOfferRewardMessage::Write); лишние 3 байта сдвигали хвост
→ деньги-награда отображались мусором (напр. «41943 золота»). Выдача всегда была
верной.

Отложено: XP-награда (до системы уровней); награда письмом при полных сумках (M7 #14).
tools/sql/quest-test-reset.sql — сброс прогресса квеста для тестов.

+++ src/AlexWoW.WorldServer/Protocol/QuestPackets.cs
@@ -151,7 +151,8 @@ public static class QuestPackets
         w.UInt32(q.Entry);
         w.CString(q.Title);
         w.CString(string.IsNullOrEmpty(q.OfferRewardText) ? q.Details : q.OfferRewardText);
-        w.UInt32(1);                 // auto_finish (Bool32) — можно завершить
+        w.UInt8(1);                  // auto_finish — u8 (НЕ u32! сверено с TrinityCore: uint8(AutoLaunched));
+                                     // лишние 3 байта сдвигали весь хвост → money врал (напр. «41943 золота»)
         w.UInt32(ClientFlags(q.QuestFlags));
         w.UInt32(q.SuggestedPlayers);
         w.UInt32(0);                 // amount_of_emotes
